refactor(ui): separate AttackSurfaceCardItem and AttackSurfaceSkeleton into own files
