+++ ui/app/(prowler)/_new-overview/components/attack-surface/index.ts
@@ -1,2 +1,4 @@
-export { AttackSurface, AttackSurfaceSkeleton } from "./attack-surface";
+export { AttackSurface } from "./attack-surface";
 export { AttackSurfaceSSR } from "./attack-surface.ssr";
+export { AttackSurfaceCardItem } from "./attack-surface-card-item";
+export { AttackSurfaceSkeleton } from "./attack-surface-skeleton";
